fix: put fluent-bit config under homebrew base

+++ uninstall_observe_mac.sh
@@ -4,15 +4,25 @@ sudo osqueryctl stop
 brew services stop telegraf
 sudo launchctl stop fluent-bit
 
-sudo rm -rf /etc/fluent-bit
-sudo rm -rf /opt/homebrew/etc/fluent-bit
-sudo rm -rf /opt/homebrew/var/fluent-bit
-sudo rm -rf /usr/local/etc/fluent-bit
-sudo rm -rf /usr/local/var/fluent-bit
-sudo rm -rf /var/osquery
-sudo rm /Library/LaunchDaemons/fluent-bit.plist
-sudo rm -rf /opt/homebrew/etc/telegraf.*
-sudo rm -rf /usr/local/brew/etc/telegraf.*
+BREW=$(which brew)
+if [ $BREW != "/opt/homebrew/bin/brew" ] && [ $BREW != "/usr/local/bin/brew" ]; then
+    echo "This script is only supported with the homebrew package manager"
+    exit
+elif [ $BREW = "/opt/homebrew/bin/brew" ]; then
+    BASE_BREW="/opt/homebrew/"
+else
+    BASE_BREW="/usr/local"
+fi
+
+if [ "$1" == "CLEAN" ];
+then
+    sudo rm -rf $BASE_BREW/fluent-bit
+    sudo rm -rf $BASE_BREW/var/fluent-bit
+    sudo rm -rf $BASE_BREW/etc/fluent-bit
+    sudo rm -rf /var/osquery
+    sudo rm /Library/LaunchDaemons/fluent-bit.plist
+    sudo rm -rf $BASE_BREW/etc/telegraf.*
+fi
 
 brew uninstall osquery
 brew uninstall telegraf
